use inheritance to dry up code
